refactor: dto 요청, 응답 분리
